# specs/001-md/contracts/api.yaml
openapi: 3.0.3
info:
  title: MOF数据t-SNE交互式可视化 API
  description: 材料科学研究用MOF数据t-SNE降维和可视化Web应用API
  version: 1.0.0
  contact:
    name: MOF Visualization Team
    email: team@mof-viz.com
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /api/visualizations/{pipeline_id}:
    get:
      summary: 获取可视化数据
      description: 获取t-SNE降维后的可视化数据，包括坐标和配置信息
      operationId: get_visualization_data
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 可视化数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisualizationData'
        '404':
          $ref: '#/components/responses/NotFound'
        '425':
          description: 处理尚未完成
  /api/visualizations/{pipeline_id}/config:
    put:
      summary: 更新可视化配置
      description: 更新可视化的外观配置（颜色、样式等）
      operationId: update_visualization_config
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisualizationConfig'
      responses:
        '200':
          description: 配置更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisualizationConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  # 导出API
  /api/visualizations/{pipeline_id}/export:
    post:
      summary: 导出可视化图像
      description: 将当前可视化导出为指定格式的图像文件
      operationId: export_visualization
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: 导出成功，返回文件下载链接
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/downloads/{file_id}:
    get:
      summary: 下载导出文件
      description: 下载已生成的导出文件
      operationId: download_file
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 文件内容
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'

  # WebSocket连接
...
components:
  schemas:
...
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 请求是否成功
        timestamp:
          type: string
          format: date-time
          description: 响应时间
        request_id:
          type: string
          format: uuid
          description: 请求唯一标识符
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            error:
              type: object
              properties:
                code:
                  type: string
                  description: 错误代码
                message:
                  type: string
                  description: 错误信息
                details:
                  type: object
                  description: 错误详情

    # 数据相关Schema
...
    CategoryInfo:
      type: object
      properties:
        category_id:
          type: integer
        category_name:
          type: string
        sample_count:
          type: integer
        color_code:
          type: string
          pattern: ^#[0-9A-Fa-f]{6}$

    # 处理相关Schema
...
    VisualizationData:
      type: object
      properties:
        visualization_id:
          type: string
          format: uuid
        pipeline_id:
          type: string
          format: uuid
        config:
          $ref: '#/components/schemas/VisualizationConfig'
        coordinates:
          type: array
          items:
            $ref: '#/components/schemas/Coordinate'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategoryInfo'
        statistics:
          $ref: '#/components/schemas/VisualizationStatistics'
        render_time_ms:
          type: integer
    Coordinate:
      type: object
      properties:
        sample_id:
          type: string
        x:
          type: number
          format: float
        y:
          type: number
          format: float
        category_id:
          type: integer
        category_name:
          type: string
        descriptive_data:
          type: object
          additionalProperties: true
        distance_to_center:
          type: number
          format: float
        local_density:
          type: number
          format: float
    VisualizationStatistics:
      type: object
      properties:
        total_samples:
          type: integer
        valid_samples:
          type: integer
        category_counts:
          type: object
          additionalProperties:
            type: integer
        x_range:
          type: object
          properties:
            min:
              type: number
              format: float
            max:
              type: number
              format: float
        y_range:
          type: object
          properties:
            min:
              type: number
              format: float
            max:
              type: number
              format: float
        density_info:
          type: object
          properties:
            high_density_regions:
              type: integer
            low_density_regions:
              type: integer
            average_density:
              type: number
              format: float
    VisualizationConfig:
      type: object
      properties:
        chart_title:
          type: string
          maxLength: 200
        x_axis_label:
          type: string
          maxLength: 100
        y_axis_label:
          type: string
          maxLength: 100
        width:
          type: integer
          minimum: 400
          maximum: 2000
          default: 800
        height:
          type: integer
          minimum: 300
          maximum: 1500
          default: 600
        show_legend:
          type: boolean
          default: true
        color_scheme:
          type: array
          items:
            $ref: '#/components/schemas/ColorConfig'
        marker_styles:
          type: array
          items:
            $ref: '#/components/schemas/MarkerConfig'
    ColorConfig:
      type: object
      properties:
        category_id:
          type: integer
        color_hex:
          type: string
          pattern: ^#[0-9A-Fa-f]{6}$
        opacity:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 1.0
    MarkerConfig:
      type: object
      properties:
        category_id:
          type: integer
        marker_type:
          type: string
          enum: [circle, square, triangle, diamond, cross]
        size:
          type: integer
          minimum: 1
          maximum: 20
          default: 6

    # 导出相关Schema
    ExportRequest:
      type: object
      properties:
        format:
          type: string
          enum: [png, svg, pdf]
          default: png
        width:
          type: integer
          minimum: 400
          maximum: 4000
        height:
          type: integer
          minimum: 300
          maximum: 3000
        dpi:
          type: integer
          minimum: 72
          maximum: 600
          default: 300
        background_color:
          type: string
          enum: [white, transparent, black]
          default: white
        filename:
          type: string
          maxLength: 100
          pattern: ^[a-zA-Z0-9_\-\s\.]+$
    ExportResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                file_id:
                  type: string
                filename:
                  type: string
                file_size_bytes:
                  type: integer
                download_url:
                  type: string
                  format: uri
                expires_at:
                  type: string
                  format: date-time
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: INVALID_REQUEST
                      message:
                        example: 请求参数格式错误或缺少必要参数
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - type: object
                properties:
                  error:
                    properties:
                      code:
                        example: NOT_FOUND
                      message:
                        example: 请求的资源不存在
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...